Refactor linear probe and fine-tuning: SQLite persistence, metadata defaults, and model compatibility
Major improvements to linear probe and fine-tuning infrastructure:

**Linear Probe:**
- Migrate persister from JSON files to SQLite database with WAL mode
- Add metadata-based defaults for task, target_col, and split_type
- Add validate() and status() methods for better builder UX
- Enforce validation in build() with clear error messages

**Fine-Tuning:**
- Remove dead _lora_applied flag
- Add post-LoRA validation to ensure trainable parameters exist
- Add check to prevent LoRA on nn.Identity backbone (NaiveBaseline)
- Consolidate metric computation into TaskHead methods
- Add prepare_targets(), score(), compute_metrics() to TaskHeadProtocol
- Fix exception handling (remove AttributeError from fallback catches)

**Model Compatibility:**
- Evo1: Add PEFT compatibility (config.to_dict method)
- Evo2: Add set_inference_mode() and set_train_mode() methods
- NaiveMamba: Add mamba-ssm 2.0 compatibility (mlp_cls parameter)
- Enformer: Add lora_target_modules specification
- OmniGenome: Remove hardcoded float16 dtype
- Orthrus: Fix model version name (orthrus-base-4 → orthrus-large-4)

**Documentation:**
- Update installation instructions: Python 3.12, PyTorch 2.7.1, CUDA 12.6
- Document FP8 configuration for Evo2 on older GPUs
- Update LinearProbeBuilder usage examples with metadata defaults
- Clarify Evo2 hardware requirements

**Dependencies:**
- Pin PEFT to 0.13.0 (compatibility requirement)
- Update transformers to 4.44.0, multimolecule to 0.0.7
- Add protobuf dependency

**Cleanup:**
- Remove deprecated linear probe scripts (by_embname.py, by_modelname.py)
- Remove obsolete orthrus_lp scripts

+++ mrna_bench/models/omnigenome.py
@@ -65,7 +65,7 @@ class OmniGenome(EmbeddingModel):
             trust_remote_code=True,
             cache_dir=get_model_weights_path(),
             token_dropout=False,
-        ).to(device, dtype=torch.float16)
+        ).to(device)
 
     def _forward_chunks(
         self,
